RL_Agent/MADDPG/requirements.txt
absl-py==2.1.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
coverage==7.4.4
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.19.1
filelock==3.13.3
flatbuffers==24.3.7
fonttools==4.49.0
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-pasta==0.2.0
grpcio==1.62.1
gym==0.10.5
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.10.0
idna==3.6
iniconfig==2.0.0
install==1.3.5
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpickle==3.0.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.0.5
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
# Editable Git install with no remote (multiagent==0.0.1)
-e /Users/anirudh/Downloads/multiagent-particle-envs-master
munch==4.0.0
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
networkx==3.2.1
numpy==1.26.4
numpy-stl==3.1.1
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyglet==2.0.14
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-utils==3.8.2
pytz==2024.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
sacred==0.8.5
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.1.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
termcolor==2.4.0
threadpoolctl==3.4.0
tinycss2==1.2.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
yarg==0.1.9
